+++ Changelog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Apply setHeader to control cache in web engine which allow to control web page caching across all browsers. Completed on 2024-05-28
 - Release simple MariaDB database client engine into sqlmanager. Completed on 2024-06-04
 - Define and standardized database query returning data format in `mariadb.js`. Completed on 2024-06-05
+- Add `ischema` method into `clsMariaDB` at `mariadb.js` to check is empty database. Completed on 2024-06-05
 
 ### Changed